requirements.txt
fastapi
uvicorn
pydantic
pandas
xlrd
python-dotenv
fuzzywuzzy
spacy
numpy
